app/requirements.txt
python-dotenv
requests
Flask
chromadb 
tiktoken 
langchain
langchain-community 
langchain-core
boto3